# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Təhvil renovation project coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects available to the signed-in participant
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a renovation project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{projectId}/dashboard:
    get:
      operationId: getProjectDashboard
      tags: [projects]
      summary: Get the project overview and action items
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
...
  /projects/{projectId}/milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List project milestones
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a project milestone
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Milestone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
...
  /projects/{projectId}/payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List the manual project payment schedule
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Payment schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentItem"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Add a manual payment schedule item
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentItem"
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        propertyType: { type: string }
        city: { type: string }
        projectType: { type: string }
        plannedStartDate: { type: ["string", "null"], format: date }
        plannedCompletionDate: { type: ["string", "null"], format: date }
        budget: { type: ["number", "null"] }
        approvedBudget: { type: number }
        approvedChanges: { type: number }
        progress: { type: integer, minimum: 0, maximum: 100 }
        status: { type: string }
        scopeVersion: { type: integer }
        scopeStatus: { type: string, enum: [draft, pending_approval, approved] }
        archived: { type: boolean }
        participantRole: { type: string, enum: [owner, contractor, viewer] }
        contractorName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, name, propertyType, city, projectType, approvedBudget, approvedChanges, progress, status, scopeVersion, scopeStatus, archived, participantRole, createdAt]
    ProjectInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        propertyType: { type: string, enum: [apartment, house, office, commercial, other] }
        city: { type: string, minLength: 1, maxLength: 100 }
        projectType: { type: string, enum: [full, partial, kitchen, bathroom, electrical, plumbing, flooring, painting, other] }
        plannedStartDate: { type: ["string", "null"], format: date }
        plannedCompletionDate: { type: ["string", "null"], format: date }
        budget: { type: ["number", "null"], minimum: 0 }
      required: [name, propertyType, city, projectType]
    Dashboard:
      type: object
      properties:
        project: { $ref: "#/components/schemas/Project" }
        roomCount: { type: integer }
        includedCount: { type: integer }
        excludedCount: { type: integer }
        pendingChanges: { type: integer }
        awaitingHandover: { type: integer }
        nextPayment: { type: ["number", "null"] }
        actionItems:
          type: array
          items: { $ref: "#/components/schemas/ActionItem" }
        milestones:
          type: array
          items: { $ref: "#/components/schemas/Milestone" }
        recentActivity:
          type: array
          items: { $ref: "#/components/schemas/TimelineEvent" }
      required: [project, roomCount, includedCount, excludedCount, pendingChanges, awaitingHandover, actionItems, milestones, recentActivity]
    ActionItem:
      type: object
      properties:
        id: { type: string }
        label: { type: string }
        detail: { type: string }
        kind: { type: string, enum: [change, handover, payment, scope] }
        href: { type: string }
      required: [id, label, detail, kind, href]
...
    Milestone:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        plannedCompletionDate: { type: ["string", "null"], format: date }
        status: { type: string, enum: [planned, in_progress, submitted_for_handover, revision_requested, accepted] }
        note: { type: ["string", "null"] }
        media:
          type: array
          items: { $ref: "#/components/schemas/MediaItem" }
        acceptedAt: { type: ["string", "null"], format: date-time }
      required: [id, projectId, title, status, media]
    MilestoneInput:
      type: object
      properties:
        title: { type: string, minLength: 2, maxLength: 120 }
        description: { type: ["string", "null"], maxLength: 1000 }
        plannedCompletionDate: { type: ["string", "null"], format: date }
      required: [title]
    MediaItem:
      type: object
      properties:
        id: { type: string }
        objectPath: { type: string }
        originalName: { type: string }
        contentType: { type: string }
        size: { type: integer }
        uploadedAt: { type: string, format: date-time }
      required: [id, objectPath, originalName, contentType, size, uploadedAt]
...
    PaymentItem:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        title: { type: string }
        amount: { type: number }
        currency: { type: string, enum: [AZN] }
        dueCondition: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"], format: date }
        status: { type: string, enum: [planned, due, marked_sent, confirmed_received, disputed] }
        note: { type: ["string", "null"] }
        media:
          type: array
          items: { $ref: "#/components/schemas/MediaItem" }
        updatedAt: { type: string, format: date-time }
      required: [id, projectId, title, amount, currency, status, media, updatedAt]
    PaymentInput:
      type: object
      properties:
        title: { type: string, minLength: 2, maxLength: 160 }
        amount: { type: number, minimum: 0 }
        dueCondition: { type: ["string", "null"], maxLength: 250 }
        dueDate: { type: ["string", "null"], format: date }
      required: [title, amount]
...
    TimelineEvent:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        action: { type: string }
        detail: { type: string }
        actorRole: { type: string, enum: [owner, contractor, viewer] }
        createdAt: { type: string, format: date-time }
      required: [id, projectId, action, detail, actorRole, createdAt]
...